git-svn-id: https://svn.code.sf.net/p/java-ml/code/trunk@168 3162d2cb-3e1f-0410-96b0-914c58786302

+++ net/sf/javaml/distance/RBFKernel.java
@@ -31,7 +31,7 @@ public class RBFKernel implements DistanceMeasure {
     private double gamma = 0.01;
 
     public RBFKernel() {
-        this(0.01f);
+        this(0.01);
     }
 
     /**
@@ -39,7 +39,7 @@ public class RBFKernel implements DistanceMeasure {
      * 
      * @param gamma
      */
-    public RBFKernel(float gamma) {
+    public RBFKernel(double gamma) {
         this.gamma = gamma;
     }
